+++ src/Hardware/DisplayST7735.h
@@ -26,5 +26,5 @@ class ST7735
 
     void Initialize();
     void SetAddressWindow(uint16_t x0, uint16_t y0, uint16_t x1, uint16_t y1);
-    void Present(uint16_t* buffer); //send full framebuffer to display
+    void Present(uint8_t* buffer); //send full framebuffer to display
 };
